refactor: 전역적으로 쓰이는 상수 관련 파일을 constants 폴더 하위로 이동
- RacingCarConstants.js 추가
- RacingErrorMessage.js 추가
- RacingMessage.js 추가
