+++ src/constants/RacingCarConstants.js
@@ -1,0 +1,7 @@
+const RacingcarConstants = {
+    MOVE_CONDITION: 4,
+    MOVE_FORWARD: 1,
+    MAX_NAME_LENGTH: 5,
+};
+
+export default RacingcarConstants;
