+++ src/constants/RacingErrorMessage.js
@@ -1,0 +1,9 @@
+const ERROR_MESSAGE = {
+    IS_EMPTY:'[ERROR] 공백이 입력되었습니다.',
+    CAR_NAME_LENGTH:'[ERROR] 이름 글자수가 5자를 초과하였습니다.',
+    IS_NUMBER:'[ERROR] 숫자가 아닙니다.',
+    IS_POSITIVE_NUMBER:'[ERROR] 0 이하의 숫자 입니다.',
+    IS_INTEGER:'[ERROR] 정수가 아닙니다.',
+};
+
+export default ERROR_MESSAGE;
