+++ src/constants/RacingMessage.js
@@ -1,0 +1,8 @@
+const MESSAGE = {
+    READ_CAR_NAME: '경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n',
+    READ_ATTEMPTS: '시도할 횟수는 몇 회인가요?\n',
+    RACING_RESULT_TITLE: '\n실행 결과',
+    WINNER_RESULT_TITLE: '최종 우승자:',
+};
+
+export default MESSAGE;
